libvdisk: Fix fcntl locking on 64-bit platforms
- The fcntl() wrapper assumed that its third argument fit in an unsigned
int, which broke fcntl locking operations (which take struct flock *) on
64-bit platforms.  Use unsigned long instead.

Debugged-by: Mike Glazer <mglazer@andrew.cmu.edu>

+++ libvdisk/libvdisk.c
@@ -366,9 +366,9 @@ int dup2(int oldfd, int newfd)
 
 int fcntl(int fd, int cmd, ...)
 {
-	unsigned arg;
+	unsigned long arg;
 	int ret, err;
-	get_last_arg(cmd, unsigned, arg);
+	get_last_arg(cmd, unsigned long, arg);
 	ret=fcntl_real(fd, cmd, arg);
 	err=errno;
 	debug("fcntl %d on %d => %d", cmd, fd, ret);
